OptimalBath

Documentation for OptimalBath.

OptimalBath.AverageBottomSource

OptimalBath.BathymetryHandlingReconstruction

OptimalBath.DryFrontTerm

OptimalBath.InnerCentralDifferenceFlux

OptimalBath.L2

OptimalBath.NoRegularization

OptimalBath.OuterCentralDifferenceFlux

OptimalBath.PrimalSWEProblem

OptimalBath.Regularization

OptimalBath.SoftL1

OptimalBath.VolumeFluxesSolver

OptimalBath.build_solver

OptimalBath.compute_objective

OptimalBath.compute_objective_and_gradient

OptimalBath.compute_objective_and_gradient!

OptimalBath.compute_Δx

OptimalBath.create_callback

OptimalBath.depth_cutoff

OptimalBath.desingularize

OptimalBath.flux_jacobian_spatial_derivative

OptimalBath.get_bathymetry

OptimalBath.get_grid

OptimalBath.initial_state

OptimalBath.objective_density

OptimalBath.objective_density_gradient

OptimalBath.primal_solver

OptimalBath.regularization_gradient!

OptimalBath.solve_adjoint

OptimalBath.solve_primal

OptimalBath.AverageBottomSource — Type

AverageBottomSource <: AdjointBottomSource

A bottom source balanced with OuterCentralDifferenceFlux when using NoReconstruction. Uses the central difference of the average bottom topography in the left and right cells.`

OptimalBath.BathymetryHandlingReconstruction — Type

BathymetryHandlingReconstruction()

A linear reconstruction similar to VolumeFluxes.LinearReconstruction, but caps heights and momentum to zero instead of adjusting the slope.

OptimalBath.DryFrontTerm — Type

DryFrontTerm()

A bottom topography source term corresponding to BathymetryHandlingReconstruction. Similar to the one used in VolumeFluxes, except that it balances the fluxes also at dry fronts, where the water height goes to zero.

OptimalBath.InnerCentralDifferenceFlux — Type

InnerCentralDifferenceFlux <: FluxJacobianDivergence

Approximates the flux Jacobian divergence ∇ · df/dU^T using the left and right reconstruction of the center state.

InnerCentralDifferenceFlux is not balanced

This flux Jacobian divergence is not balanced against the CentralUpwind numerical flux.

OptimalBath.L2 — Type

L2()

L2 regularization, divided by the number of elements in β to make it independent of the size of the bathymetry.

Example

julia> L2()(ones(4))
1.0

OptimalBath.NoRegularization — Type

NoRegularization()

OptimalBath.OuterCentralDifferenceFlux — Type

OuterCentralDifferenceFlux <: FluxJacobianDivergence

Approximates the flux Jacobian divergence ∇ · df/dU^T using the left and right average states. Used with NoReconstruction.

OuterCentralDifferenceFlux is not balanced

This flux Jacobian divergence is not balanced against the CentralUpwind numerical flux.

OptimalBath.PrimalSWEProblem — Type

PrimalSWEProblem(U0, T, grid; initial_bathymetry=zeros(grid.N))
PrimalSWEProblem(N, U0, T; initial_bathymetry=zeros(N + 1), domain=[0.0 1.0])

A data structure representing the primal shallow water equations parameters.

OptimalBath.Regularization — Type

Regularization

Defines the building cost of modifying the bathymetry. It is defined as vector space of regularization terms, so that we can easily make linear combinations of regularization terms.

Example

julia> regularization = 0.1 * L2() + 0.5 * SoftL1(10.0)
0.1 * L2() + 0.5 * SoftL1(10.0)
julia> typeof(regularization)
RegularizationSum{ScaledRegularization{L2, Float64}, ScaledRegularization{SoftL1, Float64}}

OptimalBath.SoftL1 — Type

SoftL1(α)

A smooth approximation to the L1 regularization, divided by the number of elements in β. α controls the sharpness of the approximation.

OptimalBath.VolumeFluxesSolver — Type

VolumeFluxesSolver(spec::SolverSpec, float_type)
VolumeFluxesSolver(problem::PrimalSWEProblem, options::SolverOptions, float_type)
VolumeFluxesSolver(problem::PrimalSWEProblem, reconstruction, timestepper, bathymetry_source, float_type)

A wrapper around VolumeFluxes's SWE solver.

OptimalBath.build_solver — Function

build_solver(spec::SolverSpec[, float_type::Type=Float64])

OptimalBath.compute_objective — Function

compute_objective(U::States{S, Depth}, t, Δx, β, objectives::Objectives, timestepper::Timestepper) where S
compute_objective(U::States{S, Depth}, t, Δx, β, objectives::Objectives, timestepper::Type{<:Timestepper}) where S

OptimalBath.compute_objective_and_gradient — Function

compute_objective_and_gradient(β, primal_swe_problem::PrimalSWEProblem, data::GradientData, ::GradientType)

OptimalBath.compute_objective_and_gradient! — Function

compute_objective_and_gradient!(G, β, primal_swe_problem::PrimalSWEProblem, data::GradientData, gradient_type::GradientType)

Assigns the gradient of the objective function with respect to the parameters β into G, and returns the objective function value.

OptimalBath.compute_Δx — Function

compute_Δx(solver::PrimalSWESolver)

Computes the spatial discretization step size Δx for the given solver.

OptimalBath.create_callback — Function

create_callback(f, solver::PrimalSWESolver)

Creates a callback function for the given solver that applies the function f(U_n::States, t_n, Δt) at each time step during the primal solve.

OptimalBath.depth_cutoff — Function

depth_cutoff(solver::PrimalSWESolver)

Returns the depth for which the momentum is regularized to 0.

OptimalBath.desingularize — Function

desingularize(h, p, solver::PrimalSWESolver)

Desingularizes the velocity p/h uing solver's desingularization strategy.

OptimalBath.flux_jacobian_spatial_derivative — Function

flux_jacobian_spatial_derivative(U_right_left, U_left_center, U_right_center, U_left_right, Δx, dir, ::FluxJacobianDivergence, ::ContinuousAdjointSWE)

Approximates the derivative of the flux Jacobian along dir.

Arguments

U_right_left: Right reconstruction of the left state

U_left_center: Left reconstruction of the center state

U_right_center: Right reconstruction of the center state

U_left_right: Left reconstruction of the right state

Δx: Grid spacing in the direction of interest

dir: Direction of differentiation

::FluxJacobianDivergence: Type of flux Jacobian divergence to use (for dispatch)

::ContinuousAdjointSWE: Type indicator for dispatch

OptimalBath.get_bathymetry — Function

get_bathymetry(solver::PrimalSWESolver)

OptimalBath.get_grid — Function

get_grid(solver::PrimalSWESolver)

OptimalBath.initial_state — Function

initial_state(solver::PrimalSWESolver)

OptimalBath.objective_density — Function

objective_density(obj::Objective, U::States{S, Depth, T, N, A}, I...) where {S, T, N, A}
objective_density(::Objective, U)

OptimalBath.objective_density_gradient — Function

objective_density_gradient(obj::Objective, U::States{S, Depth, T, N, A}, I...) where {S, T, N, A}
objective_density_gradient(obj::Objective, U)

OptimalBath.primal_solver — Function

primal_solver(da::AdjointSWE)

OptimalBath.regularization_gradient! — Method

regularization_gradient!(g, β, regularization)

Computes the gradient of the regularization term with respect to the bathymetry β and stores it in g.

Example

regularization = 0.1 * L2() + 0.5 * SoftL1(10.0)
0.1 * L2() + 0.5 * SoftL1(10.0)
julia> β = 1.0:4.0;
julia> g = similar(β);
julia> regularization_gradient!(g, β, regularization)
4-element Vector{Float64}:
 0.1749886505328244
 0.2249999994847116
 0.2749999999999766
 0.325

OptimalBath.solve_adjoint — Function

solve_adjoint(Λ_end, U::AverageDepthStates, objectives::Objectives, b, t, Δx, da::AdjointSWE)
solve_adjoint(Λ0, Ul::LeftDepthStates, Ur::RightDepthStates, objectives::Objectives, b, t, Δx, da::AdjointSWE)

OptimalBath.solve_primal — Function

solve_primal(solver::PrimalSWESolver, δb[, callback])

Solves the primal shallow water equations with the provided solver given the change δb in bathymetry, returning the solution U, time points t, and spatial points x.