OptimalBath

Documentation for OptimalBath.

OptimalBath.BathymetryHandlingReconstruction

OptimalBath.DryFrontTerm

OptimalBath.ForwardADGradient

OptimalBath.PrimalSWEProblem

OptimalBath.VolumeFluxesSolver

OptimalBath.adjoint_solver

OptimalBath.build_solver

OptimalBath.compute_flux_source

OptimalBath.compute_gradient!

OptimalBath.compute_objective

OptimalBath.compute_objective_and_gradient

OptimalBath.compute_objective_and_gradient!

OptimalBath.compute_Δx

OptimalBath.create_callback

OptimalBath.get_bathymetry

OptimalBath.get_gradient

OptimalBath.get_objective

OptimalBath.gradient!

OptimalBath.initial_state

OptimalBath.objective_density

OptimalBath.objective_density_gradient

OptimalBath.solve_adjoint

OptimalBath.solve_primal

OptimalBath.update_and_get_bathymetry!

OptimalBath.BathymetryHandlingReconstruction — Type

BathymetryHandlingReconstruction()

A linear reconstruction similar to VolumeFluxes.LinearReconstruction, but caps heights and momentum to zero instead of adjusting the slope.

OptimalBath.DryFrontTerm — Type

DryFrontTerm()

A bottom topography source term corresponding to BathymetryHandlingReconstruction. Similar to the one used in VolumeFluxes, except that it balances the fluxes also at dry fronts, where the water height goes to zero.

OptimalBath.ForwardADGradient — Type

ForwardADGradient(bathymetry::AbstractArray, β::AbstractArray)

A gradient computation strategy using ForwardDiff for automatic differentiation. Stores two preallocated buffers: one for the bathymetry and one for the gradient results. The bathymetry starts as a copy of the initial bathymetry, and is updated during gradient computations.

OptimalBath.PrimalSWEProblem — Type

PrimalSWEProblem(N, U0, T; initial_bathymetry=zeros(N + 1), domain=[0.0 1.0])

A data structure representing the primal shallow water equations parameters.

OptimalBath.VolumeFluxesSolver — Type

VolumeFluxesSolver(spec::SolverSpec, float_type)
VolumeFluxesSolver(problem::PrimalSWEProblem, options::SolverOptions, float_type)
VolumeFluxesSolver(problem::PrimalSWEProblem, reconstruction, timestepper, bathymetry_source, float_type)

A wrapper around VolumeFluxes's SWE solver.

OptimalBath.adjoint_solver — Function

adjoint_solver(primal_swe_problem::PrimalSWEProblem, ag::AdjointGradient) -> AdjointSolver

OptimalBath.build_solver — Function

build_solver(spec::SolverSpec[, float_type::Type=Float64])

OptimalBath.compute_flux_source — Method

compute_flux_source(Λ, Ul, Ur, Δx)

Computes the source due to the reformulation of the adjoint pde to conservative form.

OptimalBath.compute_gradient! — Function

compute_gradient!(G, Λ, U, t, Δx, objectives::Objectives, ag::AdjointGradient)
compute_gradient!(G, Λ, Ul, Ur, t, Δx, objectives::Objectives, ag::AdjointGradient)

OptimalBath.compute_objective — Function

compute_objective(U::States{S, Depth, T, N, A}, t, x, β, objectives::Objectives, timestepper::Timestepper) where {S, T, N, A}
compute_objective(U::States{S, Depth, T, N, A}, t, x, β, objectives::Objectives, timestepper::Type{<:Timestepper}) where {S, T, N, A}

OptimalBath.compute_objective_and_gradient — Function

compute_objective_and_gradient(β, primal_swe_problem::PrimalSWEProblem, data::GradientData, ::GradientType)

OptimalBath.compute_objective_and_gradient! — Function

compute_objective_and_gradient!(G, β, primal_swe_problem::PrimalSWEProblem, data::GradientData, gradient_type::GradientType)

Assigns the gradient of the objective function with respect to the parameters β into G, and returns the objective function value.

OptimalBath.compute_Δx — Function

compute_Δx(solver::PrimalSWESolver)

Computes the spatial discretization step size Δx for the given solver.

OptimalBath.create_callback — Function

create_callback(f, solver::PrimalSWESolver)

Creates a callback function for the given solver that applies the function f(U_n::States, t_n, Δt) at each time step during the primal solve.

OptimalBath.get_bathymetry — Function

get_bathymetry(solver::PrimalSWESolver)

OptimalBath.get_gradient — Function

get_gradient(ad::ADGradient)

OptimalBath.get_objective — Function

get_objective(ad::ADGradient)

OptimalBath.gradient! — Function

gradient!(ad::ADGradient, f, β)

OptimalBath.initial_state — Function

initial_state(solver::PrimalSWESolver)

OptimalBath.objective_density — Function

objective_density(obj::Objective, U::States{S, Depth, T, N, A}, I...) where {S, T, N, A}
objective_density(::Objective, U)

OptimalBath.objective_density_gradient — Function

objective_density_gradient(obj::Objective, U::States{S, Depth, T, N, A}, I...) where {S, T, N, A}
objective_density_gradient(obj::Objective, U)

OptimalBath.solve_adjoint — Function

solve_adjoint(Λ_end, U::AverageDepthStates, objectives::Objectives, b, t, Δx, da::AdjointSWE)
solve_adjoint(Λ0, Ul::LeftDepthStates, Ur::RightDepthStates, objectives::Objectives, b, t, Δx, da::AdjointSWE)

OptimalBath.solve_primal — Function

solve_primal(solver::PrimalSWESolver, δb[, callback])

Solves the primal shallow water equations with the provided solver given the change δb in bathymetry, returning the solution U, time points t, and spatial points x.

OptimalBath.update_and_get_bathymetry! — Function

update_and_get_bathymetry!(ad::ADGradient, swe_problem::PrimalSWEProblem, indices, β)