OptimalBath

Documentation for OptimalBath.

OptimalBath.ForwardADGradient

OptimalBath.SinFVMPrimalSWEProblem

OptimalBath.compute_flux_source

OptimalBath.compute_objective

OptimalBath.compute_objective_and_gradient

OptimalBath.compute_objective_and_gradient!

OptimalBath.objective_density

OptimalBath.objective_density_gradient

OptimalBath.solve_primal

OptimalBath.ForwardADGradient — Type

ForwardADGradient(bathymetry::AbstractArray, β::AbstractArray)

A gradient computation strategy using ForwardDiff for automatic differentiation. Stores two preallocated buffers: one for the bathymetry and one for the gradient results. The bathymetry starts as a copy of the initial bathymetry, and is updated during gradient computations.

OptimalBath.SinFVMPrimalSWEProblem — Type

SinFVMPrimalSWEProblem

Wrapper around SinFVM. Contains problem definition and solver parameters.

OptimalBath.compute_flux_source — Method

compute_flux_source(Λ, Ul, Ur, Δx)

Computes the source due to the reformulation of the adjoint pde to conservative form.

OptimalBath.compute_objective — Function

compute_objective(U, t, x, β, gradient_data::GradientData)

OptimalBath.compute_objective_and_gradient — Function

compute_objective_and_gradient(β, primal_swe_problem::PrimalSWEProblem, data::GradientData, ::GradientType)

OptimalBath.compute_objective_and_gradient! — Function

compute_objective_and_gradient!(G, β, primal_swe_problem::PrimalSWEProblem, data::GradientData, gradient_type::GradientType)

Assigns the gradient of the objective function with respect to the parameters β into G, and returns the objective function value.

OptimalBath.objective_density — Function

objective_density(::Objective, U)

OptimalBath.objective_density_gradient — Function

objective_density_gradient(obj::Objective, U)

OptimalBath.solve_primal — Function

solve_primal(primal_swe_problem::PrimalSWEProblem, bathymetry)

Solves the primal shallow water equations for a given primal_swe_problem and bathymetry, returning the solution U, time points t, and spatial points x.